error catching when parsing settings from file
catch any errors that may happen when parsing the settings.cfg file:
parse errors, type errors, etc.

Also changed element colours to uint8_t instead of int

+++ App/Source/ChemVis/Chemical.h
@@ -64,7 +64,7 @@ namespace ChemVis {
 
 		// Atom colour handling
 		static Core::Color GetAtomColor(int type);
-		static void SetAtomColors(std::vector<int> colours);
+		static void SetAtomColors(std::vector<uint8_t> colours);
 
 		// Atom Sybmol handling
 		static std::string GetAtomSymbol(int type);
@@ -86,7 +86,7 @@ namespace ChemVis {
 		static BondsInfo ParseBonds(Core::json data);
 		static ChemicalInfo ParseInfo(Core::json data);
 
-		static std::vector<int> s_ElementColors; // RGB
+		static std::vector<uint8_t> s_ElementColors; // RGB
 		static std::vector<std::string> s_ElementSymbols;
 	};
